녹음게시판 목록 조회 쿼리 수정

See merge request s09-webmobile1-sub2/S09P12B106!223

+++ backend/src/main/java/com/yukgaejang/voss/domain/recordboard/service/dto/response/RecordDetailResponse.java
@@ -19,11 +19,11 @@ public class RecordDetailResponse {
     private String originalFileName;
     private String savedFileName;
     private LocalDateTime createdAt;
-    private Long likes;
+    private Integer likes;
     private boolean isLiked;
     private String likeMembers;
 
-    public RecordDetailResponse(Record record, String originalFileName, String savedFileName, Long likes, Integer isLiked) {
+    public RecordDetailResponse(Record record, String originalFileName, String savedFileName, Integer likes, Integer isLiked) {
         this.recordId = record.getId();
         this.description = record.getDescription();
         this.nickname = record.getMember().getNickname();
